Significant changes!
Name class renamed to Identity because it encompassed more than a simple name and "Name" was harder to find in the source.
Resolver.resolve returns false now when unable to resolve a module, instead of a module with source === false. The previous behavior was to allow modules resolved by the fallback to be cached by Needy. That's not necessary and creates extra side-effects that have to be handled.
Removed allowUnresolved option. It dependended on the resolver always returning a cached module even if the module couldn't really be resolved. It "could" still be allowed, always returning a new Object instance instead of throwing an exception, but it's complicated by fallback and the unresolved exports object wouldn't be persistant if the same unresolved module was resolved twice.